requirements.txt
django
google-generativeai
python-dotenv
reportlab
python-docx
weasyprint
PyPDF2
pdf2image
Pillow
